Step 2

+++ _result.h
@@ -1,14 +1,14 @@
 template <class Entry> struct Result
 {
     std::filesystem::path dir;
-    std::filesystem::path filename; ioqueue<Entry> * output; Result (
-    std::filesystem::path filename, ioqueue<Entry> & output) :
-    filename(filename), output(&output)
+    std::filesystem::path filename; ioqueue<Entry> * output; bool reportage; Result (
+    std::filesystem::path filename, ioqueue<Entry> & output, bool reportage = true) :
+    filename(filename), output(&output), reportage(reportage && GENERATE_REPORTS)
     {
         std::filesystem::path name = filename.stem();
         if (name.empty()) throw std::logic_error("Result::Result: " + filename.string());
         dir = filename.parent_path() / name;
-        if (std::filesystem::is_directory(dir)) { print("remove dir: ", dir);
+        if (std::filesystem::is_directory(dir) && reportage) { print("remove dir: ", dir);
             std::filesystem::remove_all(dir);
         }
     }
@@ -28,26 +28,24 @@ template <class Entry> struct Result
     void accept (Entry entry, str report_title = "", bool content = false)
     {
         output->push (entry);
-        stream("accept", report_title) << entry; if (content)
-        stream("accept", report_title + " #") << entry.title << std::endl;
+        if (reportage) stream("accept", report_title) << entry; if (content)
+        if (reportage) stream("accept", report_title + " #") << entry.title << std::endl;
     }
     void reject (Entry entry, str report_title = "", bool content = false)
     {
-        stream("reject", report_title) << entry; if (content)
-        stream("reject", report_title + " #") << entry.title << std::endl;
+        if (reportage) stream("reject", report_title) << entry; if (content)
+        if (reportage) stream("reject", report_title + " #") << entry.title << std::endl;
     }
     void report (Entry entry, str report_title = "", bool content = false)
     {
-        stream("report", report_title) << entry; if (content)
-        stream("report", report_title + " #") << entry.title << std::endl;
+        if (reportage) stream("report", report_title) << entry; if (content)
+        if (reportage) stream("report", report_title + " #") << entry.title << std::endl;
     }
 
-    void reject (str s, str report_title = "")
-    {
-        if (s != "") stream("reject", report_title) << s;
+    void reject (str s, str report_title = "") {
+        if (reportage) if (s != "") stream("reject", report_title) << s;
     }
-    void report (str s, str report_title = "")
-    {
-        if (s != "") stream("report", report_title) << s;
+    void report (str s, str report_title = "") {
+        if (reportage) if (s != "") stream("report", report_title) << s;
     }
 };
